backend/requirements.txt
requests==2.31.0
urllib3==2.0.7
python-dotenv==1.0.0
google-search-results==2.4.2
beautifulsoup4==4.12.2
lxml==4.9.3
newspaper3k==0.2.8
openai==1.12.0
cerebras-cloud-sdk==1.50.1
